refactor: journal cleanup + docs (PR1 of #799)
Three low-risk items from the journal subsystem audit:

1. Delete `server/workspace/journal/linkRewrite.ts` (4-line
   re-export shim). The two functions (rewriteWorkspaceLinks /
   rewriteMarkdownLinks) live in `server/utils/markdown.ts`;
   the shim was pure backwards-compat clutter. dailyPass.ts now
   imports them directly. test/journal/test_linkRewrite.ts keeps
   its more-thorough edge-case coverage but points at the canonical
   source, with a comment noting why it stayed in the journal test
   subdir.

2. `parseState()` now logs at info-level when it discards state due
   to a `JOURNAL_STATE_VERSION` mismatch. The reset itself was
   intentional (cheap to rebuild) but happened silently — a
   postmortem couldn't distinguish a forced re-ingest from "first
   run after install". (Smell G in plans/audit-journal-subsystem.md.)

3. `index.ts` gains an architecture-comment block at the top
   showing the three independent pipelines (daily / optimization /
   memory extractor) hung off `maybeRunJournal()`, their cadences,
   and their outputs. Mirrors the diagram in the audit plan so a
   reader doesn't have to grep across files to find the layout.
   (Smell D.)

No behaviour change. 3066 / 3066 unit tests pass.

Refs #799.
